+++ src/engine/core/src/config.rs
@@ -15,20 +15,20 @@ pub enum Addon {
     DecomposeJongseongSsang,
 }
 
-#[derive(Serialize, Deserialize, Clone, Copy)]
+#[derive(Serialize, Deserialize, Clone, Copy, Debug)]
 pub enum HotkeyBehavior {
     ToggleHangul,
     ToHangul,
     ToEnglish,
 }
 
-#[derive(Serialize, Deserialize, Clone, Copy)]
+#[derive(Serialize, Deserialize, Clone, Copy, Debug)]
 pub enum HotkeyResult {
     Consume,
     Bypass,
 }
 
-#[derive(Serialize, Deserialize, Clone, Copy)]
+#[derive(Serialize, Deserialize, Clone, Copy, Debug)]
 pub struct Hotkey {
     behavior: HotkeyBehavior,
     result: HotkeyResult,
@@ -138,6 +138,7 @@ impl Config {
                 dir.list_config_files("layouts")
                     .into_iter()
                     .find_map(|layout| {
+                        log::info!("Try Read custom layout: {}...", layout.display());
                         if layout.file_stem()?.to_str()? == raw.layout {
                             Some(Layout::from_items(
                                 serde_yaml::from_reader(std::fs::File::open(layout).ok()?).ok()?,
@@ -152,10 +153,13 @@ impl Config {
                     ($($name:expr),+) => {
                         match raw.layout.as_str() {
                             $(
-                                $name => Layout::load_from(include_str!(concat!(concat!("../data/", $name), ".yaml"))).expect("Load builtin layout"),
+                                $name => Layout::load_from(include_str!(concat!(concat!("../data/", $name), ".yaml"))).unwrap_or_else(|_| {
+                                    log::error!("Can't load builtin layout {} fallback to empty layout", $name);
+                                    Layout::default()
+                                }),
                             )+
                             other => {
-                                eprintln!("Can't find layout {}", other);
+                                log::error!("Can't find layout {} fallback to empty layout", other);
                                 Layout::default()
                             }
                         }
@@ -173,7 +177,16 @@ impl Config {
 
         let raw = dir
             .find_config_file("config.yaml")
-            .and_then(|config| serde_yaml::from_reader(std::fs::File::open(config).ok()?).ok())
+            .and_then(|config| {
+                log::info!("Found config file: {}", config.display());
+                match serde_yaml::from_reader(std::fs::File::open(config).ok()?) {
+                    Ok(config) => Some(config),
+                    Err(err) => {
+                        log::error!("Read config error: {}", err);
+                        None
+                    }
+                }
+            })
             .unwrap_or_default();
 
         Some(Self::from_raw_config(raw, Some(dir)))