+++ src/frontends/gtk2/src/immodule.h
@@ -2,6 +2,7 @@
 #include <gdk/gdk.h>
 #include <gtk/gtk.h>
 #include <gtk/gtkimmodule.h>
+#include <kime_engine.h>
 
 void register_module(GTypeModule *module);
 GType get_kime_ty();