+++ src/frontends/xim/src/main.rs
@@ -13,7 +13,6 @@ fn main() -> Result<(), ServerError> {
     if args.contains(["-h", "--help"]) {
         println!("-h or --help: show help");
         println!("-v or --version: show version");
-        println!("--verbose: more verbose log");
         return Ok(());
     }
 
@@ -22,17 +21,12 @@ fn main() -> Result<(), ServerError> {
         return Ok(());
     }
 
-    let mut log_level = if cfg!(debug_assertions) {
-        log::LevelFilter::Trace
-    } else {
-        log::LevelFilter::Info
-    };
-
-    if args.contains("--verbose") {
-        log_level = log::LevelFilter::Trace;
-    }
-
-    simplelog::SimpleLogger::init(log_level, simplelog::ConfigBuilder::new().build()).ok();
+    assert!(
+        kime_engine_cffi::check_api_version(),
+        "Engine version mismatched"
+    );
+    kime_log::enable_logger_with_env();
+    kime_engine_cffi::enable_logger_with_env();
 
     log::info!("Start xim server version: {}", env!("CARGO_PKG_VERSION"));